requirements.txt
# База данных
sqlalchemy==1.4.42
pymysql==1.0.3
aiomysql==0.1.1
python-dotenv==1.0.1

# Веб-сервер
fastapi==0.96.0
uvicorn==0.22.0
starlette==0.28.0
pydantic==1.10.9
pydantic-settings==2.7.1
email_validator==2.1.1

# gRPC
grpcio==1.54.2
grpcio-tools==1.54.2
protobuf==4.23.3

# Рекомендательные модели
rectools[lightfm]==0.4.0
lightfm==1.17
threadpoolctl==3.1.0
apscheduler==3.10.1
pytz==2023.3
implicit==0.7.2

# Научные библиотеки
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
scipy==1.10.1

# Утилиты
tqdm==4.65.0
loguru==0.7.0
networkx==3.1
requests==2.31.0
PyYAML==6.0.1
httpx==0.27.0

# Тестирование
pytest==7.3.1
pytest-asyncio==0.21.0

python-dotenv~=1.0.1
pandas~=2.2.3
protobuf~=5.29.3
grpcio~=1.70.0
uvicorn~=0.34.0
numpy~=1.26.4
threadpoolctl~=3.5.0
RecTools~=0.12.0
pytest~=8.2.0
h11~=0.14.0
pip~=25.0
attrs~=23.2.0
build~=1.2.2.post1
wheel~=0.41.2
cryptography~=42.0.5
Jinja2~=3.1.4
keyring~=25.6.0
colorama~=0.4.6
typing_extensions~=4.12.2
sniffio~=1.3.1
setuptools~=68.2.0
charset-normalizer~=3.3.2
cffi~=1.16.0
mypy~=1.10.0
packaging~=24.0
platformdirs~=4.3.6
click~=8.1.7
joblib~=1.4.2
MarkupSafe~=2.1.5
python-dateutil~=2.9.0.post0
certifi~=2024.2.2
pydantic_core~=2.27.2
websockets~=12.0
anyio~=3.7.1
orjson~=3.10.10
urllib3~=2.2.1
httptools~=0.6.1
watchfiles~=0.21.0
six~=1.16.0
greenlet~=3.0.3
annotated-types~=0.6.0
typeguard~=4.4.1
idna~=3.7
importlib_metadata~=8.5.0
APScheduler~=3.11.0